test: setup for e2e env
