+++ e2e/tests/wallet-enable.test.ts
@@ -1,47 +1,27 @@
 import { DefaultTestEnv } from '../helpers';
 
-jest.setTimeout(40 * 1000);
+DefaultTestEnv.setupTest({ initWalletWithDefaults: true });
 
 describe('Enable wallet', function () {
-  let env: DefaultTestEnv;
-
-  beforeEach(async () => {
-    env = new DefaultTestEnv({ initWalletWithDefaults: true });
-    await env.init();
-  });
-
-  afterEach(async () => {
-    await env.dispose();
-  });
-
   test('should get the nickname when approved', async () => {
-    const page = await env.context.newPage();
-    await page.goto(env.defaultE2eData.localServerUrl);
-
-    const ckb = await env.getInjectedCkb();
-
     const enableTask = ckb.request({ method: 'wallet_enable' });
 
-    await env.approveForEnable();
+    const notificationPage = await testEnv.getNotificationPage();
+    await notificationPage.getByRole('button', { name: 'Connect' }).click();
 
     const res = await enableTask;
-    expect(res.nickname).toBe(env.defaultE2eData.nickname);
+    expect(res.nickname).toBe(testEnv.defaultE2eData.nickname);
 
     // should not ask again when already approved
     const res2 = await ckb.request({ method: 'wallet_enable' });
-    expect(res2.nickname).toBe(env.defaultE2eData.nickname);
+    expect(res2.nickname).toBe(testEnv.defaultE2eData.nickname);
   });
 
   test('should throw when user reject approval', async () => {
-    const page = await env.context.newPage();
-    await page.goto(env.defaultE2eData.localServerUrl);
-
-    const ckb = await env.getInjectedCkb();
-
     const enableTask = ckb.request({ method: 'wallet_enable' });
 
-    const notificationPage = await env.getNotificationPage();
-    await notificationPage.getByRole('button', { name: /cancel/i }).click();
+    const notificationPage = await testEnv.getNotificationPage();
+    await notificationPage.getByRole('button', { name: 'Cancel' }).click();
 
     await expect(enableTask).rejects.toThrowError(/reject/);
   });
